server UPDATE replace all cbs with ln2 cb
